Restructure files, add webpack config for CSS, gitignore everything under static/frontend and more README improvements
