+++ machinevision/frontend/src/components/streaming/WebcamStream/WebcamStream.js
