+++ machinevision/frontend/src/index.js
@@ -1,1 +1,2 @@
 import App from './App';
+import './styles.css';
