+++ machinevision/frontend/src/views/DashboardView/DashboardView.css
@@ -1,0 +1,8 @@
+.dashboard-container {
+  background-color: #cfe8fc;
+  /* Placeholder background image for testing */
+  background: url('../../assets/background.jpg');
+  background-repeat: no-repeat;
+  background-size: cover;
+  height: 100vh;
+}
